requirements.txt
Flask==3.1.0
Flask_SocketIO==5.5.1
marshmallow==3.19.0
psycopg2==2.9.10
